@@ -5,10 +5,10 @@
     "keywords": ["laravel", "framework"],
     "license": "MIT",
     "require": {
-        "php": "^8.2",
+        "php": "^8.3",
         "bepsvpt/secure-headers": "^7.5",
         "itsgoingd/clockwork": "^5.2",
-        "laravel/framework": "^11.9",
+        "laravel/framework": "^12.8",
         "laravel/sanctum": "^4.0",
         "laravel/slack-notification-channel": "^3.2",
         "laravel/tinker": "^2.9",
