@@ -4,7 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Exceptions\TranslatableException;
 use App\Models\User;
+use Closure;
 use Illuminate\Foundation\Bus\DispatchesJobs;
+use Illuminate\Http\Request;
 
 class AuthenticatedController extends Controller
 {
@@ -19,7 +21,7 @@ class AuthenticatedController extends Controller
 
     public function __construct()
     {
-        $this->middleware(function ($request, $next) {
+        $this->middleware(function (Request $request, Closure $next) {
             // For this controller, we require an authenticated user of type User
             if (!$request->user() instanceof User) {
                 throw new TranslatableException(
