@@ -2,8 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Exceptions\TranslatableException;
+use Closure;
 use Illuminate\Foundation\Bus\DispatchesJobs;
+use Illuminate\Http\Request;
 
 class UnauthenticatedController extends Controller
 {
@@ -11,7 +12,7 @@ class UnauthenticatedController extends Controller
 
     public function __construct()
     {
-        $this->middleware(function ($request, $next) {
+        $this->middleware(function (Request $request, Closure $next) {
             // Validate if there is no user authenticated
             if (!is_null($request->user())) {
                 throw new TranslatableException(
