@@ -38,9 +38,10 @@ return [
 
         'database' => [
             'driver'          => 'database',
-            'table'           => env('DB_CACHE_TABLE', 'cache'),
             'connection'      => env('DB_CACHE_CONNECTION'),
+            'table'           => env('DB_CACHE_TABLE', 'cache'),
             'lock_connection' => env('DB_CACHE_LOCK_CONNECTION'),
+            'lock_table'      => env('DB_CACHE_LOCK_TABLE'),
         ],
 
         'file' => [
