@@ -30,8 +30,10 @@ return [
     'disks' => [
         'local' => [
             'driver' => 'local',
-            'root'   => storage_path('app'),
+            'root'   => storage_path('app/private'),
+            'serve'  => true,
             'throw'  => false,
+            'report' => false,
         ],
 
         'public' => [
@@ -40,6 +42,7 @@ return [
             'url'        => env('APP_URL') . '/storage',
             'visibility' => 'public',
             'throw'      => false,
+            'report'     => false,
         ],
 
         's3' => [
@@ -52,6 +55,7 @@ return [
             'endpoint'                => env('AWS_ENDPOINT'),
             'use_path_style_endpoint' => env('AWS_USE_PATH_STYLE_ENDPOINT', false),
             'throw'                   => false,
+            'report'                  => false,
         ],
     ],
 
