@@ -37,10 +37,10 @@ return [
     'mailers' => [
         'smtp' => [
             'transport'    => 'smtp',
+            'scheme'       => env('MAIL_SCHEME'),
             'url'          => env('MAIL_URL'),
             'host'         => env('MAIL_HOST', '127.0.0.1'),
             'port'         => env('MAIL_PORT', 2525),
-            'encryption'   => env('MAIL_ENCRYPTION', 'tls'),
             'username'     => env('MAIL_USERNAME'),
             'password'     => env('MAIL_PASSWORD'),
             'timeout'      => null,
@@ -83,6 +83,7 @@ return [
                 'smtp',
                 'log',
             ],
+            'retry_after' => 60,
         ],
 
         'roundrobin' => [
@@ -91,6 +92,7 @@ return [
                 'ses',
                 'postmark',
             ],
+            'retry_after' => 60,
         ],
     ],
 
